Create module for navbar functions
